Add flag to replace `f` with a single-character sneak
It's called `sneakReplacesF` and is nice because it basically makes `f` work across newlines.
I had to pull a few interfaces and a base class out from motion.ts to avoid a circular import.
Closes #4012
